style: Move lib functions out of barrel file
