+++ lib/dateFormatter.js
@@ -1,0 +1,5 @@
+export const dateFormatter = new Intl.DateTimeFormat("en-US", {
+  month: "short",
+  day: "numeric",
+  year: "numeric",
+});
