+++ lib/findFirstImage.js
@@ -1,0 +1,9 @@
+import * as prismicH from "@prismicio/helpers";
+
+export const findFirstImage = (slices) => {
+  const imageSlice = slices.find((slice) => slice.slice_type === "image");
+
+  if (imageSlice && prismicH.isFilled.image(imageSlice.primary.image)) {
+    return imageSlice.primary.image;
+  }
+};
